rebase done

+++ Numerics.cpp
@@ -7,7 +7,7 @@ using namespace irc;
 
 void	irc::sendNumeric(int fd, std::string msg) {
 	int bytes_send = send(fd, msg.c_str(), msg.size(), 0);
-	if (bytes_send < msg.size()) {
+	if ((size_t)bytes_send < msg.size()) {
 		std::cerr << "Error while sending message to fd = " << fd << std::endl;
 		// TODO close connection
 	}
@@ -134,7 +134,6 @@ void	irc::sendNumeric(int fd, std::string msg) {
 /* 502 */ std::string	irc::ErrUsersDontMatch () { return (" :Cannot change mode for tohers users"); }
 
 
-// For user parameter :
 void	irc::numericReply(int num_nb, User* user, std::vector<std::string>& s_params) {
 	std::string					msg;
 	char		 				s_num_nb[4];
